feat: fixing DI
