+++ EsportsApi.Core/Interfaces/IValorantService.cs
@@ -1,0 +1,9 @@
+using EsportsApi.Core.Models;
+
+namespace EsportsApi.Core.Interfaces;
+
+public interface IValorantService
+{
+    List<ValorantEvent> GetLiveEvents();
+    List<ValorantMatch> GetLiveMatches(string eventUrl);
+}
